requirements.txt
argon2-cffi==16.3.0
Django==2.0
django-bootstrap4==0.0.4
django-ratelimit==1.1.0
djangorestframework==3.7.3
praw==5.2.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.3
regex==2017.12.12
unicodedata2==10.0.0.post2
